Updates to make more useable for others, with build service.
Update testborders to use standard sizes. The test tool also takes
a -quit option to exit immediately, which is more useful if you're
not running it in the desktop (eg under Pyromaniac).

Update the Makefiles to use -apcs and create directories as needed.
They're also typed now, so that they are more recognisable.